refactoring

+++ Backend/Backend/Controllers/LocationsController.cs
@@ -8,57 +8,98 @@ namespace WepAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class LocationController : ControllerBase
+    public class LocationsController : ControllerBase
     {
         private readonly ILocationService _locationService;
         private readonly ILocationMapper _locationMapper;
-        public LocationController(ILocationService locationService, ILocationMapper locationMapper)
+        public LocationsController(ILocationService locationService, ILocationMapper locationMapper)
         {
             _locationService = locationService;
             _locationMapper = locationMapper;
         }
 
+        // GET api/Locations
         [HttpGet]
-        [Route("locations")]
+        [Route("")]
         public IActionResult GetAll()
         {
             IEnumerable<LocationDTO> locationsDTO = _locationService.GetAll();
             return Ok(locationsDTO);
         }
 
+        // GET api/Locations/1
         [HttpGet]
-        [Route("locations/{id}")]
+        [Route("{id}")]
         public IActionResult GetById(int id)
         {
             LocationDTO locationDTO = _locationService.GetById(id);
+
+            if (locationDTO == null)
+            {
+                return NotFound();
+            }
+
             return Ok(locationDTO);
         }
 
+        // POST api/Locations
         [HttpPost]
-        [Route("locations")]
+        [Route("")]
         public IActionResult Add(LocationDTO locationDTO)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             _locationService.Add(locationDTO);
             _locationService.SaveLocation();
             return Ok(locationDTO);
         }
 
+        // PUT api/Locations/1
         [HttpPut]
-        [Route("locations/{id}")]
+        [Route("{id}")]
         public IActionResult Update(int id, LocationDTO locationDTO)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if (id != locationDTO.LocationId)
+            {
+                return BadRequest();
+            }
+
+            if (!LocationExists(id))
+            {
+                return NotFound();
+            }
+
             _locationService.Update(id, locationDTO);
             _locationService.SaveLocation();
             return Ok(locationDTO);
         }
 
+        // DELETE api/Location/1
         [HttpDelete]
-        [Route("locations/{id}")]
+        [Route("{id}")]
         public IActionResult Delete(int id)
         {
+            if (!LocationExists(id))
+            {
+                return NotFound();
+            }
             _locationService.Delete(id);
             _locationService.SaveLocation();
             return Ok();
+        }
+
+
+        private bool LocationExists(int id)
+        {
+            LocationDTO locationDTO = _locationService.GetById(id);
+            return locationDTO != null;
         }
     }
 }
